requirements.txt
# WebSocket client library
websocket-client==1.3.2

# HTTP requests library
requests==2.27.1

# Environment variable management
python-dotenv==0.19.2

# Dependency injection
injector==0.19.0

# Logging
loguru==0.5.3

# Alpaca API
alpaca-trade-api==3.2.0

# Interactive Brokers API
ib_insync==0.9.72

# Utilities for pandas, numpy (useful for data manipulation)
pandas==1.3.5
numpy==1.21.5

# Date and time handling
pytz==2021.3

# Argument parsing
argparse==1.4.0

# TensorFlow dependencies
tensorflow-probability==0.24.0
tf-agents==0.5.0

# Additional dependencies for TensorFlow
absl-py>=0.6.1
decorator
gast>=0.3.2
protobuf>=3.11.3

# Linting and code quality (optional)
flake8==4.0.1

# Testing (optional)
pytest==7.0.1
